Fixed Newtonsoft serialization via migration to Controller-CQRS

+++ backend/Backend.Application/DTOs/ShipmentDto.cs
@@ -6,6 +6,6 @@ public record ShipmentDto
 {
     public required int Id { get; set; }
     public required ShipmentType ShipmentType { get; init; }
-    public required int Cost { get; init; }
+    public required decimal Cost { get; init; }
     public required int Days { get; init; }
 }
